+++ faebryk/importers/netlist/kicad/__init__.py
@@ -1,5 +1,5 @@
 # This file is part of the faebryk project
 # SPDX-License-Identifier: MIT
 
+import faebryk.importers.netlist.kicad.schematic_kicad
 import faebryk.importers.netlist.kicad.netlist_kicad
-import faebryk.importers.netlist.kicad.sexp
