+++ faebryk/libs/__init__.py
@@ -4,3 +4,5 @@
 
 import faebryk.libs.util
 import faebryk.libs.exceptions
+import faebryk.libs.pycodegen
+import faebryk.libs.algorithm
