+++ faebryk/libs/algorithm.py
@@ -1,0 +1,27 @@
+class ufds:
+    parent_node = {}
+    rank = {}
+
+    def make_set(self, u):
+        for i in u:
+            self.parent_node[i] = i
+            self.rank[i] = 0
+
+    def op_find(self, k):
+        if self.parent_node[k] != k:
+            self.parent_node[k] = self.op_find(self.parent_node[k])
+        return self.parent_node[k]
+
+    def op_union(self, a, b):
+        x = self.op_find(a)
+        y = self.op_find(b)
+
+        if x == y:
+            return
+        if self.rank[x] > self.rank[y]:
+            self.parent_node[y] = x
+        elif self.rank[x] < self.rank[y]:
+            self.parent_node[x] = y
+        else:
+            self.parent_node[x] = y
+            self.rank[y] = self.rank[y] + 1
